+++ rollup.minify.js
@@ -1,16 +1,17 @@
 import buble from 'rollup-plugin-buble';
+import uglify from 'rollup-plugin-uglify';
 
 export default {
 	input: 'src/esm/index.js',
 	name: 'phoneFns',
-	external: ['callingCodes'],
 	output: {
 		format: 'umd',
-		file: 'phone-fns.js'
+		file: 'phone-fns.min.js'
 	},
 	plugins: [
 		buble({
 			exclude: ['callingCodes.js']
-		})
+		}),
+		uglify()
 	]
 };