+++ src/esm/index.js
@@ -4,3 +4,5 @@ export { default as isValid } from '../isValid/index';
 export { default as identical } from '../identical/index';
 export { default as format } from '../format/index';
 export { default as find } from '../find/index';
+export { default as findLocal } from '../findLocal/index';
+export { default as getCode } from '../getCode/index';