+++ src/getCode/index.js
@@ -1,0 +1,3 @@
+import callingCodes from '../callingCodes.js';
+
+export default country => callingCodes[country.toLowerCase()];