+++ changelog.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## v0.2.0
+
+> - Tweaks to build process
+> - Added ability to call functional methods individually youc an do so using the following:
+>   - `phone-fns/format`
+>   - `phone-fns/find`
+>   - `phone-fns/breakdown`
+>   - `phone-fns/identical`
+>   - `phone-fns/isValid`
+>   - `phone-fns/uglify`
+>   - `phone-fns/getCode`
+>   - `phone-fns/findLocal`
+>   - `phone-fns/callingCodes` - Gives back the full list of country calling codes
+>   - You can still use just `phone-fns` to get back all of the methods in an `object`
+> - Added a minified version `phone-fns.min.js` **the module points to this version by default**
+> - Renamed the file `phone-fns.umd.js` to just `phone-fns.js`
+> - Added Country Calling Codes from the [Countries Repo](https://github.com/mledoze/countries)
+> - Added `findLocal` function which returns the country names associated with the passed in code
+> - Added `getCode` function which returns the country calling code from the `json`
+> - Removed country code dashing, since it's not really a thing
+
 ## v0.1.0
 
 > - Initial Public Release